backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
pydantic-settings>=2.2.1
httpx>=0.27.0
python-multipart>=0.0.9
pyjwt>=2.8.0
pdf2image>=1.17.0
Pillow>=10.3.0
openai>=1.14.0
pytest>=8.0.0
